Refactor the custom logger for aplicate the observer pattern.
